SLM-62 Use modules for API types
