+++ server/@types/prisonRegisterApiClient/index.d.ts
@@ -1,0 +1,6 @@
+declare module 'prisonRegisterApiClient' {
+  import { components } from '../prisonRegisterApi'
+
+  export type ErrorResponse = components['schemas']['ErrorResponse']
+  export type PrisonDto = components['schemas']['PrisonDto']
+}
