--- server/@types/prisonRegisterApiClientTypes.ts
@@ -1,4 +1,0 @@
-import { components } from './prisonRegisterApi'
-
-export type ErrorResponse = components['schemas']['ErrorResponse']
-export type PrisonDto = components['schemas']['PrisonDto']
