+++ server/routes/barcode/contacts/ChooseContactController.test.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from 'express'
 import { SessionData } from 'express-session'
 import type { PrisonAddress } from 'prisonTypes'
+import type { Contact } from 'sendLegalMailApiClient'
 import PrisonRegisterService from '../../../services/prison/PrisonRegisterService'
 import ContactService from '../../../services/contacts/ContactService'
 import ChooseContactController from './ChooseContactController'
-import { Contact } from '../../../@types/sendLegalMailApiClientTypes'
 
 jest.mock('../../../config')
